Move CachedActivityData to domain/model and remove duplicate
The CachedActivityData class was duplicated - once inside
CalculateActivityDataUseCase and once as a separate file in
presentation/state. This moves the canonical definition to
domain/model where it belongs (contains only domain types)
and updates all imports accordingly.
