--- feature/habits/presentation/src/commonMain/kotlin/space/be1ski/vibits/feature/habits/presentation/state/CachedActivityData.kt
@@ -1,14 +1,0 @@
-package space.be1ski.vibits.feature.habits.presentation.state
-
-import space.be1ski.vibits.feature.habits.domain.model.ActivityWeekData
-import space.be1ski.vibits.feature.habits.domain.model.HabitsConfigEntry
-import space.be1ski.vibits.feature.habits.domain.model.SuccessRateData
-
-/**
- * Cached activity data for a specific range and mode.
- */
-data class CachedActivityData(
-  val weekData: ActivityWeekData,
-  val configTimeline: List<HabitsConfigEntry>,
-  val successRate: SuccessRateData?,
-)
