+++ feature/habits/presentation/src/commonMain/kotlin/space/be1ski/vibits/feature/habits/presentation/state/HabitsState.kt
@@ -1,8 +1,10 @@
 package space.be1ski.vibits.feature.habits.presentation.state
+
 import kotlinx.datetime.LocalDate
 import space.be1ski.vibits.feature.habits.domain.HabitTag
 import space.be1ski.vibits.feature.habits.domain.IsSelected
 import space.be1ski.vibits.feature.habits.domain.model.ActivityWeek
+import space.be1ski.vibits.feature.habits.domain.model.CachedActivityData
 import space.be1ski.vibits.feature.habits.domain.model.ContributionDay
 import space.be1ski.vibits.feature.habits.domain.model.DailyMemoInfo
 import space.be1ski.vibits.feature.habits.domain.model.HabitConfig
